expansion_calc/python_src/requirements.txt
futsu
numpy
pillow
pytesseract
